requirements.txt
fastapi
uvicorn
pillow
onnxruntime
torchvision
numpy
python-multipart